requirements.txt
click
PyYAML
py_eureka_client==0.9.8
requests
prettytable